Fix permission issue

+++ src/inject/inject.js
@@ -9,20 +9,22 @@ chrome.extension.sendMessage({}, function(response) {
         $('.votecell .vote').each(function() {
           var questionId = $(this).find('input[name=_id_]').val();
           if (questionId) {
-            $(getLikeButtonTemplate(questionId)).appendTo($(this)).on('click', function() {
-              if (this.classList.contains('active')) {
-                this.classList.remove("active");
-                deleteQuestionById(questionId);
-              } else {
-                this.classList.add("active");
+            if ($('.sofl-like-button#' + questionId).length == 0) {
+              $(getLikeButtonTemplate(questionId)).appendTo($(this)).on('click', function() {
+                if (this.classList.contains('active')) {
+                  this.classList.remove("active");
+                  deleteQuestionById(questionId);
+                } else {
+                  this.classList.add("active");
 
-                storeQuestion({
-                  questionId: questionId,
-                  questionHeader: $('#question-header h1 a').text(),
-                  questionType: $(this).parents('.question#question').length > 0 ? 'question' : 'answer'
-                });
-              }
-            });
+                  storeQuestion({
+                    questionId: questionId,
+                    questionHeader: $('#question-header h1 a').text(),
+                    questionType: $(this).parents('.question#question').length > 0 ? 'question' : 'answer'
+                  });
+                }
+              });
+            }
           }
         });
       }
